Docker Toolbox Wireup
Running communications with the UI and API using an nginx reverse proxy. This setup is configured for use with a local docker machine cooked with Docker Toolbox.

This could be made more universal with careful environment configuration and separate dockerfiles per use case, but in the interest of time I'm going to move on to data concerns.